refactor(plugins): split the plugins command module by responsibility
plugins.rs had grown to 800+ lines across three concerns. Now a module
folder: mod.rs (InstalledPlugin + the Tauri commands), manifest.rs
(parsing and validation, the security boundary), store.rs (disk
operations: scan, folder/package install, asset reads, uninstall), and
tests.rs per the sync/git/tests.rs precedent for large test modules.
No behavior change.
